# docs/swagger.yaml
definitions:
...
  disability_system_backend_internal_modules_incapacidades_dto.SubirDocumentoRequest:
    properties:
      formato:
        type: string
      id_incapacidad:
        type: integer
      nombre:
        type: string
      tipo:
        type: string
      url:
        type: string
    required:
    - formato
    - id_incapacidad
    - nombre
    - tipo
    - url
    type: object
...
  disability_system_backend_internal_modules_incapacidades_dto.ValidarDocumentoRequest:
    properties:
      comentario:
        type: string
      estado:
        type: string
    required:
    - estado
    type: object
...
info:
  contact: {}
paths:
...
  /documentos/{id}:
    delete:
      consumes:
      - application/json
      description: Elimina un documento del sistema
      parameters:
      - description: ID del documento
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Eliminar documento
      tags:
      - documentos
  /documentos/{id}/validar:
    patch:
      consumes:
      - application/json
      description: Valida o rechaza un documento
      parameters:
      - description: ID del documento
        in: path
        name: id
        required: true
        type: integer
      - description: Estado de validación
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/disability_system_backend_internal_modules_incapacidades_dto.ValidarDocumentoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Validar documento
      tags:
      - documentos
...
  /incapacidades/{id}/documentos:
    get:
      consumes:
      - application/json
      description: Lista los documentos de una incapacidad
      parameters:
      - description: ID de la incapacidad
        in: path
        name: id
        required: true
        type: integer
      - description: ID de la incapacidad
        in: query
        name: id_incapacidad
        required: true
        type: integer
      - description: Filtrar por estado
        in: query
        name: estado
        type: string
      - description: Filtrar por tipo
        in: query
        name: tipo
        type: string
      - default: 1
        description: Página
        in: query
        name: page
        type: integer
      - default: 20
        description: Límite
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Listar documentos
      tags:
      - documentos
    post:
      consumes:
      - application/json
      description: Registra un documento asociado a una incapacidad
      parameters:
      - description: Datos del documento
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/disability_system_backend_internal_modules_incapacidades_dto.SubirDocumentoRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Subir documento
      tags:
      - documentos
  /incapacidades/{id}/documentos/url:
    post:
      consumes:
      - application/json
      description: Genera una URL prefirmada para subir documentos a R2
      parameters:
      - description: ID de la incapacidad
        in: path
        name: id
        required: true
        type: integer
      - description: Datos del archivo (nombre, formato, tipo)
        in: body
        name: request
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Generar URL prefirmada para upload
      tags:
      - documentos
...
  /incapacidades/{id}/historial:
    get:
      consumes:
      - application/json
      description: Lista el historial de cambios de una incapacidad
      parameters:
      - description: ID de la incapacidad
        in: path
        name: id
        required: true
        type: integer
      - description: ID de la incapacidad
        in: query
        name: id_incapacidad
        type: integer
      - description: Filtrar por tipo de evento
        in: query
        name: id_tipo_historial
        type: integer
      - default: 1
        description: Página
        in: query
        name: page
        type: integer
      - default: 20
        description: Límite
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Listar historial de incapacidad
      tags:
      - incapacidades
...
swagger: "2.0"